refactor(chat-components): AB#131731 dedup layouts via shared helpers
Extracted BaseLayoutProps, MatchedPlugins, and MessageFocusTarget to
src/layouts/shared/. Unified focus-target id prefix so the existing
moveFocusToMessageFocusTarget util works under c26. Moved layouts up to
src/layouts/. Replaced :global(.user) with [data-source="user"]. Dropped
narration comments and redundant narrowing ternaries.
