+++ src/layouts/shared/index.ts
@@ -1,0 +1,3 @@
+export type { BaseLayoutProps } from "./types";
+export { MatchedPlugins } from "./MatchedPlugins";
+export { MessageFocusTarget } from "./MessageFocusTarget";
